backend/requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
PyMySQL==1.1.0
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.1
uvicorn==0.23.2
zipp==3.16.2
